requirements.txt
aiohttp==3.6.2
captcha==0.3
python-dateutil==2.8.1
Unidecode==1.1.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.2
bs4==0.0.1
captcha==0.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dhooks==1.1.4
discord.py==1.5.1
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
Pillow==7.2.0
psutil==5.7.2
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
yarl==1.6.0
columnar==1.3.1
pymongo==3.11.0
dnspython==2.0.0
python-dateutil==2.8.1